backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
PyMuPDF==1.24.1
sentence-transformers==2.2.2
transformers==4.34.0
tokenizers>=0.14,<0.15
faiss-cpu==1.12.0
python-dotenv==1.0.0
tqdm==4.66.1
numpy>=1.25,<3.0
pandas==2.2.2
google-generativeai==0.8.5
protobuf>=5.26.1  